third_party/WebKit/LayoutTests/device_orientation/orientation/no-synchronous-events.html

<!DOCTYPE html>
<html>
<body>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../resources/device-orientation-helpers.js"></script>
<script>
'use strict';

async_test(test => {
  assertTestRunner();
  var orientationData = generateOrientationData(1.1, 2.2, 3.3, true);

  setMockOrientation(orientationData);

  var hasAddEventListenerReturned = false;
  window.addEventListener('deviceorientation', test.step_func_done(() => {
    assert_true(hasAddEventListenerReturned);
  }));
  hasAddEventListenerReturned = true;
}, 'Tests that events are never fired synchronously from a call to window.addEventListener().');
</script>
</body>
</html>
